Состояние

+++ shbizadv/src/main/java/org/example/bizarreadventure/repository/UserRepository.java
@@ -1,6 +1,5 @@
 package org.example.bizarreadventure.repository;
 
-
 import org.example.bizarreadventure.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
@@ -8,9 +7,9 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface UserRepository extends JpaRepository<User, Long> {
+public interface UserRepository extends JpaRepository<User, Integer> {
     boolean existsByEmail(String email);
     boolean existsByLogin(String login);
     User findByLogin(String login);
-    Optional<User> findById(Long id);
+    Optional<User> findById(int id);
 }
